try to get some vertical space back
